Added multi controller support and adjusted dragon controller
Added:
- support for multiple dragons each controller by a different controller
- cooldown on dragon fireball to prevent spamming
- added placeholder dragon model

+++ FireballProject/Assets/Scripts/FireballController.cs
@@ -6,7 +6,7 @@ public class FireballController : MonoBehaviour
 {
     // damage dealt to entities hit by the explosion
     [SerializeField]
-    private float explosionDamage = 0f;
+    private int explosionDamage = 0;
 
     // effect to spawn when the fireball explodes
     [SerializeField]
